feat: modify exisiting logic for revocation in didSignedDocumentStatus verifier
