+++ src/types/error.ts
@@ -57,6 +57,7 @@ export enum OpenAttestationDidSignedDocumentStatusCode {
   CORRESPONDING_PROOF_MISSING = 6,
   DID_MISSING = 7,
   UNRECOGNIZED_DOCUMENT = 8,
+  UNRECOGNIZED_REVOCATION_TYPE = 9
 }
 export enum OpenAttestationDidCode {
   SKIPPED = 0,
