+++ src/verifiers/documentStatus/types.ts
@@ -1,0 +1,18 @@
+import { Reason } from "../../types/error";
+
+interface ValidRevocationStatus {
+  revoked: false;
+  address: string;
+}
+
+export interface InvalidRevocationStatus {
+  revoked: true;
+  address: string;
+  reason: Reason;
+}
+
+export interface IdentityRevocationStatus {
+  revoked: boolean;
+}
+
+export type RevocationStatus = ValidRevocationStatus | InvalidRevocationStatus | IdentityRevocationStatus;
